pyproject.toml
[tool.poetry]
name = "test_work_v2"
version = "0.1.0"
description = ""
authors = ["Viktor Nikolaev <viktornicolaev1995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
Pillow = "^9.3.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
serializers = "^0.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
